+++ src/components/Header/Header.css.ts
@@ -2,5 +2,5 @@ import { atoms } from '../../strum-design-system/sprinkles.css';
 
 export const headerStyle = atoms({
   backgroundColor: 'primary',
-  padding: 2,
+  paddingY: 8,
 });